Rename payment request
